+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## nextbone-routing log
 
+### [v1.0.0] - 2024-06-24
+ - Upgrade slick-router to v3.0.0
+ - Remove ability to elEvent decorator listen to nextbone events
+ - Export `elEvent` decorator as `eventHandler`
+ - Support custom property hooks 
+
 ### [v0.9.0] - 2019-07-03
 
 - Use slick-router instead of cherrytreex